+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 - New %7Bcourseid%7D as an alternative to {courseid} for use with encoded URLs.
 - New %7Buserid%7D as an alternative to {userid} for use with encoded URLs.
 - New {coursestartdate} tag.
+- New {courseenddate} tag.
 
 ## [0.4.3] - 2018-03-30
 ### Added