+++ CHANGELOG.md
@@ -7,7 +7,7 @@ All notable changes to this project will be documented in this file.
 - New %7Buserid%7D as an alternative to {userid} for use with encoded URLs.
 - New {coursestartdate} tag.
 - New {courseenddate} tag.
-
+- New {coursecompletiondate} tag.
 ## [0.4.3] - 2018-03-30
 ### Added
 - Support for reCAPTCHA v2 in Moodle 3.1.11+, 3.2.8+, 3.3.5+, 3.4.5+ and 3.5+.