+++ CHANGELOG.md
@@ -1,13 +1,18 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [0.4.5] - 2018-05-18
+### Added
+- New %7Bsesskey%7D tag as an alternative to {sesskey} for use with encoded URLs.
+
 ## [0.4.4] - 2018-05-08
 ### Added
-- New %7Bcourseid%7D as an alternative to {courseid} for use with encoded URLs.
-- New %7Buserid%7D as an alternative to {userid} for use with encoded URLs.
+- New %7Bcourseid%7D tag as an alternative to {courseid} for use with encoded URLs.
+- New %7Buserid%7D tag as an alternative to {userid} for use with encoded URLs.
 - New {coursestartdate} tag.
 - New {courseenddate} tag.
 - New {coursecompletiondate} tag.
+
 ## [0.4.3] - 2018-03-30
 ### Added
 - Support for reCAPTCHA v2 in Moodle 3.1.11+, 3.2.8+, 3.3.5+, 3.4.5+ and 3.5+.